tests: split pkg/testutil/testregistry
