+++ cmd/nerdctl/multi_platform_linux_test.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/containerd/nerdctl/pkg/testutil"
 	"github.com/containerd/nerdctl/pkg/testutil/nettestutil"
+	"github.com/containerd/nerdctl/pkg/testutil/testregistry"
 	"gotest.tools/v3/assert"
 )
 
@@ -56,10 +57,10 @@ func TestMultiPlatformBuildPush(t *testing.T) {
 	testutil.RequireExecPlatform(t, "linux/amd64", "linux/arm64", "linux/arm/v7")
 	base := testutil.NewBase(t)
 	tID := testutil.Identifier(t)
-	reg := newTestRegistry(base)
-	defer reg.cleanup()
+	reg := testregistry.NewPlainHTTP(base)
+	defer reg.Cleanup()
 
-	imageName := fmt.Sprintf("localhost:%d/%s:latest", reg.listenPort, tID)
+	imageName := fmt.Sprintf("localhost:%d/%s:latest", reg.ListenPort, tID)
 	defer base.Cmd("rmi", imageName).Run()
 
 	dockerfile := fmt.Sprintf(`FROM %s
@@ -79,10 +80,10 @@ func TestMultiPlatformPullPushAllPlatforms(t *testing.T) {
 	testutil.DockerIncompatible(t)
 	base := testutil.NewBase(t)
 	tID := testutil.Identifier(t)
-	reg := newTestRegistry(base)
-	defer reg.cleanup()
+	reg := testregistry.NewPlainHTTP(base)
+	defer reg.Cleanup()
 
-	pushImageName := fmt.Sprintf("localhost:%d/%s:latest", reg.listenPort, tID)
+	pushImageName := fmt.Sprintf("localhost:%d/%s:latest", reg.ListenPort, tID)
 	defer base.Cmd("rmi", pushImageName).Run()
 
 	base.Cmd("pull", "--all-platforms", testutil.AlpineImage).AssertOK()
